pyproject.toml
[tool.poetry]
name = "trust ai collection"
version = "0.1.0"
description = ""
authors = ["Ivan Kavalerov <ivan@kavalerov.net>"]

[tool.poetry.dependencies]
python = "^3.8"
gradio = "^3.14.0"
numpy = "^1.23.5"
scikit-learn = "^1.2.0"
chart-studio = "^1.1.0"
supabase = "^1.0.2"
python-dotenv = "^1.0.0"
openai = "^0.27.8"


[tool.poetry.group.dev.dependencies]
setuptools = "^67.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
